improved file structure / started re-using views cross platform
