+++ Sources/WishKit/Extensions/WishResponse+Identifiable.swift
@@ -1,0 +1,11 @@
+//
+//  WishResponse+Identifiable.swift
+//  wishkit-ios
+//
+//  Created by Martin Lasek on 9/27/23.
+//  Copyright © 2023 Martin Lasek. All rights reserved.
+//
+
+import WishKitShared
+
+extension WishResponse: Identifiable { }
