+++ Sources/WishKit/SwiftUI/SeparatorView.swift
@@ -1,6 +1,6 @@
 //
-//  SwiftUIView.swift
-//  
+//  SeparatorView.swift
+//  wishkit-ios
 //
 //  Created by Martin Lasek on 8/12/23.
 //  Copyright © 2023 Martin Lasek. All rights reserved.
